Revert "Feature/marge pipe path#8"
